Add files via upload
thermal_da.py is for the data collected by FLIR when placed in the collimator, 100s of tif files were collected for varying temperatures (30C, 40C, 50C, 60C, 70C, 80C) of the knife edge target, then a region of interest was cropped up, and the intensity values across a row of pixels were plotted (it is important to note that we are in the spatial domain), then I took the derivative of the ESF to get the LSF and then Fourier transform to get the MTF (it is important to note that we are in the frequency domain). Taking the FT is important because it provides me information about the resolution of the camera I am testing or characterizing.